fix: inferred state type incompatible with return type of factory functions. remove state generic constraints
Fixes an issue where the generics passed to the factory functions that create Sessions and Channels would not be passed to the Session and Channel types inside the function, causing an error to be thrown as the state was incompatible as it defaulted to unknown.

Removes Record constraints on state type generics.