fix(Session.test): fix tests broken by fetch api changes
Update all tests to use the 'waitForConnect' utility function instead of using a callback to listen on the 'connected' event being emitted.

Update push tests to await the Promise returned from the 'push' method.

Update tests related to headers to use the exact same 'DEFAULT_RESPONSE_HEADERS' object as the implementation code rather than copying it in the test file.

Update test related to headers to be case-insensitive for header names.